libblrs update and example autotune routine
the libblrs update fixed the recalculation issues we had found before
and the libfbc update introduced stall detection, I updated the existing
controllers to ignore stall detection but there is a default function
for this now available in its header file if you want to use it.

Op control is also currently set to run an autotuning routine for the
drivetrain, defining AUTOTUNE to 0 will prevent that from running.